admin-phanquyen

+++ app/Http/Controllers/admin/ProductController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\admin;
+
 use App\Models\Category;
 use App\Models\Product;
 use App\Http\Controllers\Controller;
@@ -16,7 +17,7 @@ class ProductController extends Controller
 {
     private $productService;
     private $categoryService;
-    public function __construct(ProductServiceInterface  $productService ,CategoryServiceInterface $categoryService)
+    public function __construct(ProductServiceInterface  $productService, CategoryServiceInterface $categoryService)
     {
         $this->productService = $productService;
         $this->categoryService = $categoryService;
@@ -28,6 +29,7 @@ class ProductController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Product::class);
         $products = Product::all();
         $categories = Category::get();
         $products = $this->productService->all($request);
@@ -45,13 +47,14 @@ class ProductController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Product::class);
         $products = Product::all();
         $categories = Category::all();
         $params = [
             'categories' => $categories,
             'products' => $products,
         ];
-        return view('admin.product.add',$params);
+        return view('admin.product.add', $params);
     }
 
     /**
@@ -64,15 +67,15 @@ class ProductController extends Controller
     {
         $data = $request->all();
         try {
-        $this->productService->create($data);
-        $notification = [
-            'message' => 'Thêm sản phẩm thành công!',
-            'alert-type' => 'success'
-        ];
-        return redirect()->route('products.index')->with($notification);
+            $this->productService->create($data);
+            $notification = [
+                'message' => 'Thêm sản phẩm thành công!',
+                'alert-type' => 'success'
+            ];
+            return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.store.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi xay ra!',
                 'alert-type' => 'error'
@@ -100,9 +103,10 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', Product::class);
         $categories = Category::get();
         $product = $this->productService->find($id);
-        return view('admin.product.edit' , compact('product','categories')) ;
+        return view('admin.product.edit', compact('product', 'categories'));
     }
 
     /**
@@ -115,8 +119,8 @@ class ProductController extends Controller
     public function update(UpdateProductRequest $request, $id)
     {
         $data = $request->all();
-        try{
-            $this->productService->update( $id, $data);
+        try {
+            $this->productService->update($id, $data);
             $notification = [
                 'message' => 'Cập nhật thành công!',
                 'alert-type' => 'success'
@@ -124,7 +128,7 @@ class ProductController extends Controller
             return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.update.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -141,7 +145,8 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        try{
+        $this->authorize('delete', Product::class);
+        try {
             $product = $this->productService->delete($id);
             $notification = [
                 'message' => 'Sản phẩm đã vào thùng rác!',
@@ -150,7 +155,7 @@ class ProductController extends Controller
             return redirect()->route('products.index')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.destroy.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -159,13 +164,16 @@ class ProductController extends Controller
         }
     }
 
-    public function trash(){
+    public function trash()
+    {
         $products = $this->productService->getTrash();
         return view('admin.product.trash', compact('products'));
     }
 
-    public function restore($id){
-        try{
+    public function restore($id)
+    {
+        $this->authorize('restore', Product::class);
+        try {
             $this->productService->restore($id);
             $notification = [
                 'message' => 'Đã khôi phục sản phẩm!',
@@ -174,7 +182,7 @@ class ProductController extends Controller
             return redirect()->route('product.trash')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.restore.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
@@ -183,8 +191,10 @@ class ProductController extends Controller
         }
     }
 
-    public function forcedelete($id){
-        try{
+    public function forcedelete($id)
+    {
+        try {
+            $this->authorize('forceDelete', Product::class);
             $this->productService->forceDelete($id);
             $notification = [
                 'message' => 'Đã xóa sản phẩm',
@@ -193,7 +203,7 @@ class ProductController extends Controller
             return redirect()->route('product.trash')->with($notification);
         } catch (\Exception $e) {
             Session::flash('error', config('define.forceDelete.error'));
-            Log::error('message:'. $e->getMessage());
+            Log::error('message:' . $e->getMessage());
             $notification = [
                 'message' => 'có lôi say ra!',
                 'alert-type' => 'error'
